Move generic XML parsing helpers to ParserHelper
