+++ NAS2D/Configuration.cpp
@@ -23,31 +23,6 @@ using namespace NAS2D;
 
 
 namespace {
-	Dictionary attributesToDictionary(const Xml::XmlElement& element)
-	{
-		Dictionary dictionary;
-		for (const auto* attribute = element.firstAttribute(); attribute; attribute = attribute->next())
-		{
-			dictionary.set(attribute->name(), attribute->value());
-		}
-		return dictionary;
-	}
-
-
-	std::map<std::string, Dictionary> subTagsToDictionaryMap(const Xml::XmlElement& element)
-	{
-		std::map<std::string, Dictionary> sections;
-		for (auto childElement = element.firstChildElement(); childElement; childElement = childElement->nextSiblingElement())
-		{
-			if (childElement->type() != Xml::XmlNode::NodeType::XML_COMMENT)
-			{
-				sections[childElement->value()] = attributesToDictionary(*childElement);
-			}
-		}
-		return sections;
-	}
-
-
 	template <auto fileFormatParser(const Xml::XmlElement& element) = subTagsToDictionaryMap>
 	auto parseXmlFileData(const std::string& xmlString, const std::string& sectionName = "", const std::string& requiredVersion = "")
 	{
